feat: Update Docker configuration and enhance student deletion logic to remove associated records

+++ src/main/java/com/example/inspire_environment/service/impl/StaffServiceImpl.java
@@ -143,6 +143,9 @@ public class StaffServiceImpl implements StaffService {
         if (!staffRepository.existsById(id)) {
             throw new ResourceNotFoundException("Staff", "id", id);
         }
+        if(id.equals(1L)){
+            throw new ConflictException("Cannot delete that admin");
+        }
         staffRepository.deleteById(id);
     }
